Disables ES tasks and refactors deploy sripts
