+++ README.md
@@ -199,34 +199,6 @@ APPLICATION_MODE=<environment> fab -H <bastion-host-domain> db.restore-snapshot
 ```
 
 
-Elastic Search data
--------------------
-
-> NB: It's recommended to [create your own copies of indices with the harvester commands](harvester/README.md#harvesting),
-> instead of loading (possibly old and stale) production snapshots.
-
-It's possible to load data into your localhost Elastic Search cluster with a few commands
-In order to do that you first need to setup with:
-
-```bash
-invoke es.setup
-```
-
-Backups are stored in a so called repository. You'll need to download the latest ES repository file to load the data.
-Ask somebody for the file and name of the latest backup repository and run:
-
-```bash
-invoke es.load-repository <repository-file>
-invoke es.restore-snapshot <repository-name>
-```
-
-This should have loaded the indices you need to make searches locally.
-Alternatively you can set the
-``POL_ELASTIC_SEARCH_HOST``, ``POL_ELASTIC_SEARCH_PROTOCOL``, ``POL_ELASTIC_SEARCH_USERNAME`` and
-``POL_SECRETS_ELASTIC_SEARCH_PASSWORD`` variables inside your ``.env`` file.
-This allows to connect your local setup to a remote development or testing cluster.
-
-
 Linting
 -------
