+++ commands/deploy/__init__.py
@@ -1,0 +1,3 @@
+from .container import prepare_builds, build, push
+from .deploy import deploy
+from .migrate import migrate
