+++ tasks.py
@@ -1,8 +1,7 @@
 from invoke import Collection
 
 from environments.surfpol import create_configuration_and_session
-from elastic.tasks import setup, create_snapshot, load_repository, restore_snapshot
-from deploy import prepare_builds, build, push, deploy, migrate
+from commands.deploy import prepare_builds, build, push, deploy, migrate
 from test import e2e_tests, harvester_tests, service_tests, elastic_search_tests, test
 from commands.postgres.invoke import import_snapshot as postgres_import_snapshot
 from commands.projects.harvester.invoke import setup_harvester, import_dataset, harvest, cleanup
@@ -14,7 +13,6 @@ environment, session = create_configuration_and_session(use_aws_default_profile=
 
 
 namespace = Collection(
-    Collection("es", setup, create_snapshot, load_repository, restore_snapshot),
     Collection("db", postgres_import_snapshot),
     Collection("srv", setup_service),
     Collection("hrv", setup_harvester, import_dataset, harvest, cleanup),
